M8.1 _interpret split into smaller modules
