+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "holoso"
 dynamic = ["version"]
-requires-python = ">=3.13"  # math.fma is 3.13+. Verification targets Python 3.14+.
+requires-python = ">=3.14"
 description = "A simple Python-to-Verilog synthesizer for numeric kernels, control systems, and DSP."
 readme = { file = "README.md", content-type = "text/markdown" }
 license = "Apache-2.0"
